feat: add session-based auth (#10)
* fix: various typing fixes to green the buld

* feat: add session-based auth

Add an endpoint to which clients can submit a delegation that allows the server to create a session for them. This lets us bootstrap classic secure-cookie based session auth, in this case provided by `iron-session` - a recommendation from @peeja that I think is delightful.

I've added auth to a few endpoints and moved the oauth system to use the session auth but have not tested anything other than the oauth flow - need to set up a database tomorrow.

* fix: remove console.log

* fix: add environment variables for staging and prod

* fix: deploy script err

* fix: env var name

* fix: re-remove account gates

* fix: prettier

* fix: various tweaks from PR review

* fix: get the typecasting right