Remove "use client" from Dialog component
As it was, `onClose` was invalid, because it wasn't serializable. `"use
client"` marks a client entry module, meaning a point in the tree where
a server component renders a client component. A server component can't
render this client component, because it can't pass it an `onClose`
value. But that's okay, it wouldn't make much sense anyway.