Force at least one data item on
Note that this turns the switches into controlled components. Normally
our form controls are uncontrolled, but here we need to have more, well,
control over them. However, we're still not using those controlling
*values* from the state in our submission. We still use the `FormData`
we get from the actual form at submission time. Thus, the change to a
controlled component is a only a local change.