fix: Tell 1Password to ignore `keyMaterial`
Without this, 1Password will prompt to save the `keyMaterial` field as a
password. 1Password is a great place to store the key material, but it
doesn't function like a password, so the automatic system doesn't really
work for us, and this in particular is not where the value is *created*
but where it's *used*.

The `data-1p-ignore` attribute is not documented, but it does appear to
work.