feat: delete reposiory snapshots and blobs from a backup
When people open a backup that is already archived, there's a button to delete the content of &mdash; or the backup itself(?) &mdash; that backup.

For this to work, @travis, my first approach was to do a bunch of INNER JOINS from the neccessary tables. But, I soon found out that there's a better way to attempt this &mdash; CTEs (Common Table Expressions) as it helps us group these queries based on the specific relationship(s), in this case, a backup id.

I did, however, include a JOIN for blobs in a snapshot &mdash; I don't think we can directly query at_blobs with a backup_id. Blobs belong to a snapshot, so I made a reference between snapshots and backup_id

Now, that we already have all the CIDs in a backup, I'm trying to remove them like you suggested, with client.remove(CID, {shards: true}), but I run into this error everytime.

```shell
Error: Agent has no proof for did:keyz6Mkesf9iTE....
```

From doing `await client.setCurrentSpace(backup?.storachaSpace)`