feat: disable identity transfer for now
Remove the whole "take control" concept entirely - it actually isn't necessary for anything other than creating a user account.

We do still need to take control of a DID in order to create a new account with the same `did:plc` - move `takeControl` there and use it when necessary, but also introduce a feature flag so we can ship before identity transfer is fully ready for the masses.

Users can now add `?identity-transfer=true` to the end of their URL to enable the identity transfer flow.