chore: set up plausible (#190)
Install [Plausible](https://plausible.io/) and configure it to user
lifecycle events.

This required moving a couple server actions to a "return a value and
redirect client side" model to give us a place to track "success"
events. I think the result is actually a bit more natural in both cases.

---------

Co-authored-by: kaf-lamed-beyt <belac335@gmail.com>